# groot: improve perf of reading slices of builtins