… go.sum
… groot/cmd/root-gen-type/testdata/small-evnt-tree-fullsplit.txt
… groot/cmd/root-gen-type/testdata/streamers.txt
… groot/rbytes/rbuffer_test.go
… groot/rbytes/sliceops.go
… groot/rbytes/sliceops_test.go
… groot/rbytes/wbuffer_test.go
… groot/rcont/array_gen.go
… groot/rdict/gen_type.go
… groot/rdict/object.go
… groot/rdict/rdict.go
… groot/rhist/graph.go
… groot/rhist/hist.go
… groot/rhist/hist_test.go
… groot/rhist/rootio_test.go
… groot/riofs/key.go
… groot/rtree/basket.go
… groot/rtree/branch.go
… groot/rtree/leaf_gen.go
… groot/rtree/read_streamers.go
… groot/rtree/rw_test.go
… groot/rtree/scanner_test.go
… groot/rtree/tchain_test.go
… groot/rtree/tree.go
… groot/rtree/tree_test.go