```
name                      old time/op    new time/op    delta
ReadTreeSliF64/0-4           660µs ± 3%     217µs ± 3%   -67.14%  (p=0.000 n=20+20)
ReadTreeSliF64/1-4           663µs ± 4%     217µs ± 4%   -67.25%  (p=0.000 n=20+20)
ReadTreeSliF64/2-4           664µs ± 4%     215µs ± 2%   -67.57%  (p=0.000 n=20+19)
ReadTreeSliF64/4-4           669µs ± 3%     217µs ± 3%   -67.52%  (p=0.000 n=19+20)
ReadTreeSliF64/8-4           665µs ± 3%     219µs ± 6%   -67.00%  (p=0.000 n=20+20)
ReadTreeSliF64/16-4          665µs ± 2%     218µs ± 3%   -67.26%  (p=0.000 n=20+20)
ReadTreeSliF64/64-4          664µs ± 3%     216µs ± 2%   -67.47%  (p=0.000 n=20+17)
ReadTreeSliF64/128-4         657µs ± 3%     216µs ± 3%   -67.12%  (p=0.000 n=18+19)
ReadTreeSliF64/512-4         659µs ± 3%     218µs ± 2%   -66.90%  (p=0.000 n=20+20)
ReadTreeSliF64/1024-4        663µs ± 3%     221µs ± 3%   -66.70%  (p=0.000 n=19+19)
ReadTreeSliF64/1048576-4     661µs ± 3%     225µs ± 9%   -65.92%  (p=0.000 n=19+20)

name                      old alloc/op   new alloc/op   delta
ReadTreeSliF64/0-4           396kB ± 0%       0kB ± 0%  -100.00%  (p=0.000 n=19+14)
ReadTreeSliF64/1-4           396kB ± 0%       0kB ± 0%  -100.00%  (p=0.000 n=19+14)
ReadTreeSliF64/2-4           396kB ± 0%       0kB ± 8%  -100.00%  (p=0.000 n=20+20)
ReadTreeSliF64/4-4           396kB ± 0%       0kB ± 0%  -100.00%  (p=0.000 n=20+17)
ReadTreeSliF64/8-4           396kB ± 0%       0kB ± 0%  -100.00%  (p=0.000 n=19+15)
ReadTreeSliF64/16-4          396kB ± 0%       0kB ± 0%  -100.00%  (p=0.000 n=20+17)
ReadTreeSliF64/64-4          396kB ± 0%       0kB ± 0%  -100.00%  (p=0.000 n=19+14)
ReadTreeSliF64/128-4         396kB ± 0%       0kB ± 7%  -100.00%  (p=0.000 n=20+18)
ReadTreeSliF64/512-4         396kB ± 0%       0kB ± 0%  -100.00%  (p=0.000 n=20+16)
ReadTreeSliF64/1024-4        396kB ± 0%       0kB ± 0%  -100.00%  (p=0.000 n=20+17)
ReadTreeSliF64/1048576-4     396kB ± 0%       0kB ± 7%  -100.00%  (p=0.000 n=20+18)

name                      old allocs/op  new allocs/op  delta
ReadTreeSliF64/0-4             987 ± 0%         0       -100.00%  (p=0.000 n=20+20)
ReadTreeSliF64/1-4             987 ± 0%         0       -100.00%  (p=0.000 n=20+20)
ReadTreeSliF64/2-4             987 ± 0%         0       -100.00%  (p=0.000 n=20+20)
ReadTreeSliF64/4-4             987 ± 0%         0       -100.00%  (p=0.000 n=20+20)
ReadTreeSliF64/8-4             987 ± 0%         0       -100.00%  (p=0.000 n=20+20)
ReadTreeSliF64/16-4            987 ± 0%         0       -100.00%  (p=0.000 n=20+20)
ReadTreeSliF64/64-4            987 ± 0%         0       -100.00%  (p=0.000 n=20+20)
ReadTreeSliF64/128-4           987 ± 0%         0       -100.00%  (p=0.000 n=20+20)
ReadTreeSliF64/512-4           987 ± 0%         0       -100.00%  (p=0.000 n=20+20)
ReadTreeSliF64/1024-4          987 ± 0%         0       -100.00%  (p=0.000 n=20+20)
ReadTreeSliF64/1048576-4       987 ± 0%         0       -100.00%  (p=0.000 n=20+20)
```