@sbinet 2020-05-05
@rmadar: this is mainly for slices of builtins, flat-trees.
other use-cases might get _some_ improvements from:

- smaller garbage collector pressure
- use of new (slightly better) decompression library.

but as can be seen from the benchmark figures above, varying the length of the slice of `float64` doesn't impact that much the time to read the tree.
this means the biggest overhead comes from actually loading the branches...