@rmadar 2020-05-05
Thanks for this! I didn't (yet) benchmark it on slices, but on float branches the improvements is of the order of 15%. I have updated https://github.com/rmadar/perf-gohep . I'll try with a TTree having slices and post the numbers.