@sbinet 2020-05-05
this might not be a hard limit:
- better code through better compiler (this happens, from time to time)
- reorganize code a bit
- chase and reduce memory allocations

tap into those idle cores:
- load branches in parallel
- load events in parallel

I am starting to play a bit with this in my `groot-pload` branch.