@rmadar 2020-05-05
> I'll try with a TTree having slices and post the numbers.

I have added [there](https://github.com/rmadar/perf-gohep) a fake data generation with 30 branches, either `float64` or `[]float64` over 100 kEvts. I tried default go-hep version and last commit for both cases. I see 20-22% improvement in both cases, so not a larger improvement for slices (their length is randomly distributed between 0 and 30).

It's not impossible that I am doing something wrong but the code looks quite simple (and the generated data look OK).