@sbinet 2020-05-06
The numbers I've shown are for a tree with a single slice.

This would still point to a performance issue loading the "multitude" of branches.