@sbinet 2020-05-06
I'll look at your cpu profile.