@sbinet 2020-05-06
Ah. And you've created your file with compression enabled (the default), whereas I've disabled it to disentangle it from decoding speed.