@rmadar 2020-05-06
> Ah. And you've created your file with compression enabled (the default), whereas I've disabled it to disentangle it from decoding speed. 

Ha interesting ...
 1. how can I disable it?
 2. if I do, comparison with ROOT is unfair I guess?

I have added ROOT numbers [here](https://github.com/rmadar/perf-gohep) for vector branches (well `array[N]`) both with `t->Branch()` and `TTreeReader` approaches. Numbers are quite interesting: 
  + `Branch`: 22ms / kEvts
  + `TTreeReader`: **8 ms / kEvts** (here, I suspect - hope - all data are not loaded, just the used one)
  + `go-hep`: 45 ms / kEvts