@sbinet 2020-05-06
have a look at my benchmarks:
https://github.com/go-hep/hep/blob/master/groot/rtree/rw_test.go#L1814..L1828

yes comparing groot-no-compression vs ROOT-with-compression would be unfair.
I was just pointing out the origin of the discrepancies b/w your perf numbers and mine.