@rmadar 2020-05-06
> `TTreeReader`: **8 ms / kEvts** (here, I suspect - hope - all data are not loaded, just the used one)

Actually, I have modified the test to sum over all vector elements of all branches (to make sure all data are effectively loaded), and the time goes from 8ms to 11ms. So `TTreeReader` is bascially twice faster than `t->Branch()`, which is (currently) twice faster than `rtree.ReadVar`. So I don't how they do, but they are quite fast (maybe they are some ideas to grasp?).